convert math.ceil() to int in Py2.7

+++ sdk/eventhub/azure-eventhub/azure/eventhub/_eventprocessor/ownership_manager.py
@@ -146,7 +146,7 @@ class OwnershipManager(object):  # pylint:disable=too-many-instance-attributes
             0 if self.owner_id in active_ownership_by_owner else 1
         )
         expected_count_per_owner = all_partition_count // owners_count
-        max_count_per_owner = math.ceil(all_partition_count / owners_count)
+        max_count_per_owner = int(math.ceil(all_partition_count / owners_count))
         # end of calculating expected count per owner
 
         to_claim = active_ownership_self
